move imagenet dataset to vision
